some small validation changes

+++ src/middlewares/verifyOtp.js
@@ -8,9 +8,12 @@ export const verifyOtp = async (req, res, next) => {
         if (!user) {
             throw new ApiError(404, 'User not found');
         }
-        const user_otp = await sql`SELECT * FROM otp WHERE user_id = ${user_id} AND otp_code = ${otp_code} AND request_type = ${request_type}`;
+        const user_otp = await sql`SELECT * FROM otp WHERE user_id = ${user_id} AND otp_code = ${otp_code} AND request_type = ${request_type} AND`;
         if (user_otp.length === 0) {
-            throw new ApiError(400, 'Invalid or expired OTP');
+            throw new ApiError(400, 'Incorrect OTP');
+        }
+        if (user_otp[0].expires_at < new Date()) {
+            throw new ApiError(400, 'OTP expired');
         }
         if (user_otp[0].otp_code === otp_code) {
             await sql`DELETE FROM otp WHERE user_id = ${user_id}`;
